+++ src/graphparser.js
@@ -53,66 +53,60 @@ module.exports = class GraphParser {
    */
   tokenize (input) {
     var self = this
-    var tokens = []
-    var matches = [self.whitespace.default]
-    var _matches = []
+    var tokens = [self.whitespace.default] /* holds matched strings */
+    var token
     var lastIndex = 0
-    var prevWhitespace = false
+    var prevWhitespace = true
     var isWhitespace = function (token) {
       return self.tokens[token].includes(self.whitespace.tokenClass)
     }
-    var m
+    var m /* RegExp match object */
     while ((m = self.tokenizer.exec(input)) !== null) {
-      _matches.push(m)
-      matches.push(m[1])
+      token = m[1] /* save matched token string */
       if (m.index !== lastIndex) {
-        self.valueError('Unrecognized token at pos ' + lastIndex + ' of ' + input)
+        self.valueError(
+          'Unrecognized token "' +
+          input.substring(lastIndex, m.index) + '" at pos ' +
+          (lastIndex) + ' of ' + input
+
+        )
       }
-      lastIndex = _matches[_matches.length - 1].index +
-                   _matches[_matches.length - 1][1].length
-    }
-    if (matches.length === 1) {
-      self.valueError('Unrecognized token at pos 0 of ' + input)
-      return []
-    }
-    for (var i = 0; i < matches.length; i++) {
-      var token = matches[i]
-      if (isWhitespace(token)) {
-        if (prevWhitespace && self.whitespace.consolidate) {
-          continue
+      if (self.whitespace.consolidate) {
+        if (isWhitespace(token)) {
+          if (prevWhitespace) {
+            continue
+          } else {
+            prevWhitespace = true
+            token = self.whitespace.default
+          }
         } else {
-          prevWhitespace = true
+          prevWhitespace = false
         }
-      } else {
-        prevWhitespace = false
       }
       tokens.push(token)
+      lastIndex = self.tokenizer.lastIndex
     }
-    if (self.whitespace.consolidate) {
-      while (isWhitespace(tokens[tokens.length - 1])) {
-        tokens.pop()
-      }
+
+    if (self.whitespace.consolidate === false ||
+        prevWhitespace === false) {
+      tokens.push(self.whitespace.default)
     }
-    tokens.push(self.whitespace.default)
-    if (lastIndex !== input.length) {
-      self.ValueError(
-        'Unrecognizable input at pos ' + lastIndex +
-          ' of ' + input
-      )
-    }
+
     return tokens
   }
 
   /**
    * @param {number} startIdx index of token to start match at
    * @param {[]} constraintVal
+   * @param {[]} tokens
    * @param {boolean} checkPrev
    * @param {boolean} checkNext
    * @param {boolean} byClass
    * @return {boolean}
    */
-  matchTokens (startIdx, constraintVal, checkPrev, checkNext, byClass) {
-    var tokens = this._input_tokens
+
+  matchTokens (startIdx, constraintVal, tokens, checkPrev, checkNext, byClass) {
+    // var tokens = this._input_tokens
     if (checkPrev && startIdx < 0) {
       return false
     }
@@ -137,7 +131,7 @@ module.exports = class GraphParser {
    * @param {number} tokenIdx
    * @return {boolean}
    */
-  matchConstraints (source, target, tokenIdx) {
+  matchConstraints (source, target, tokenIdx, tokens) {
     var self = this
     var targetEdge = self.graph.edge[source][target]
     var constraints = targetEdge['constraints']
@@ -150,40 +144,45 @@ module.exports = class GraphParser {
 
     for (var constraintType in constraints) {
       constraintVal = constraints[constraintType]
-      if (constraintType === 'prev_tokens') {
-        numTokens = self.rules[self.graph.node[target]['rule_key']].tokens.length
-        startAt = tokenIdx
-        startAt -= numTokens
-        startAt -= constraintVal.length
-        if (!self.matchTokens(startAt, constraintVal, true, false, false)) {
-          return false
-        }
-      } else if (constraintType === 'next_tokens') {
-        startAt = tokenIdx
-        if (!self.matchTokens(startAt, constraintVal, false, true, false)) {
-          return false
-        }
-      } else if (constraintType === 'prev_classes') {
-        numTokens = self.rules[self.graph.node[target]['rule_key']].tokens.length
-        startAt = tokenIdx
-        startAt -= numTokens
-        prevTokens = constraints['prev_tokens']
-        if (prevTokens) {
-          startAt -= prevTokens.length
-        }
-        startAt -= constraintVal.length
-        if (!this.matchTokens(startAt, constraintVal, true, false, true)) {
-          return false
-        }
-      } else if (constraintType === 'next_classes') {
-        startAt = tokenIdx
-        nextTokens = constraints['next_tokens']
-        if (nextTokens) {
-          startAt += nextTokens.length
-        }
-        if (!this.matchTokens(startAt, constraintVal, false, true, true)) {
-          return false
-        }
+      switch (constraintType) {
+        case 'prev_tokens':
+          numTokens = self.rules[self.graph.node[target]['rule_key']].tokens.length
+          startAt = tokenIdx
+          startAt -= numTokens
+          startAt -= constraintVal.length
+          if (!self.matchTokens(startAt, constraintVal, tokens, true, false, false)) {
+            return false
+          }
+          break
+        case 'next_tokens':
+          startAt = tokenIdx
+          if (!self.matchTokens(startAt, constraintVal, tokens, false, true, false)) {
+            return false
+          }
+          break
+        case 'prev_classes':
+          numTokens = self.rules[self.graph.node[target]['rule_key']].tokens.length
+          startAt = tokenIdx
+          startAt -= numTokens
+          prevTokens = constraints['prev_tokens']
+          if (prevTokens) {
+            startAt -= prevTokens.length
+          }
+          startAt -= constraintVal.length
+          if (!this.matchTokens(startAt, constraintVal, tokens, true, false, true)) {
+            return false
+          }
+          break
+        case 'next_classes':
+          startAt = tokenIdx
+          nextTokens = constraints['next_tokens']
+          if (nextTokens) {
+            startAt += nextTokens.length
+          }
+          if (!this.matchTokens(startAt, constraintVal, tokens, false, true, true)) {
+            return false
+          }
+          break
       }
     }
     return true
@@ -191,42 +190,40 @@ module.exports = class GraphParser {
 
   /**
   * @param {number} tokenIdx
+  * @param {[]} tokens
   * @param {boolean} matchAll
   * @return {number|number[]}
   */
-  matchAt (tokenIdx, matchAll) {
+  matchAt (tokenIdx, tokens, matchAll) {
     if (matchAll) {
       var matches = []
     }
     var self = this
-    var tokens = self._input_tokens
+    // var tokens = self._input_tokens
     var graph = self.graph
     var stack = []
     var appendChildren = function (nodeKey, tokenIdx) {
       var children = null
-      var orderedChildren = graph.node[nodeKey]['ordered_children']
-      var childKey, rulesKeys, rulesKey
+      var orderedChildren = graph.node[nodeKey]['ordered_children'] || []
+      var childKey, rulesKeys, ruleKey
 
-      if (orderedChildren) {
-        children = orderedChildren[tokens[tokenIdx]]
-        if (children) {
-          for (var i = children.length; i >= 0; i--) {
-            childKey = children[i]
-            /* stack (LIFO) from right side */
-            stack.push([childKey, nodeKey, tokenIdx])
-          }
-        } else {
-          rulesKeys = orderedChildren['__rules__']
-          if (rulesKeys) {
-            for (var j = rulesKeys.length; j >= 0; j--) {
-              rulesKey = rulesKeys[j]
-              stack.push([rulesKey, nodeKey, tokenIdx])
-            }
+      children = orderedChildren[tokens[tokenIdx]]
+      if (children) {
+        for (var i = children.length - 1; i >= 0; i--) {
+          childKey = children[i]
+          /* stack (LIFO) from right side */
+          stack.push([childKey, nodeKey, tokenIdx])
+        }
+      } else {
+        rulesKeys = orderedChildren['__rules__']
+        if (rulesKeys) {
+          for (var j = rulesKeys.length - 1; j >= 0; j--) {
+            ruleKey = rulesKeys[j]
+            stack.push([ruleKey, nodeKey, tokenIdx])
           }
         }
       }
     }
-
     appendChildren(0, tokenIdx)
 
     while (stack.length > 0) {
@@ -235,30 +232,22 @@ module.exports = class GraphParser {
       var parentKey = x[1]
       tokenIdx = x[2]
       var currNode = graph.node[nodeKey]
-
       if (currNode['accepting'] &&
-        self.matchConstraints(parentKey, nodeKey, tokenIdx)) {
+        self.matchConstraints(parentKey, nodeKey, tokenIdx, tokens)) {
         if (matchAll) {
-          matches.append(currNode['rule_key'])
+          matches.push(currNode['rule_key'])
+          continue
         } else {
           return currNode['rule_key']
         }
-        continue
       } else {
         if (tokenIdx < tokens.length - 1) {
           tokenIdx += 1
         }
         appendChildren(nodeKey, tokenIdx)
       }
-    } if (matchAll) {
-      return matches
-    } else {
-      self.ValueError(
-        'Could not match token ' + tokens[tokenIdx] +
-        ' at pos ' + tokenIdx +
-        ' in ' + tokens
-      )
     }
+    return matches
   }
 
   /**
@@ -272,10 +261,9 @@ module.exports = class GraphParser {
     var tokenIdx = 1
     var currMatchRules, prevToken, currToken, currTokenRules, tokensMatched,
       ruleKey, rule
-    self._input_tokens = tokens
     self._ruleKeys = []
     while (tokenIdx < tokens.length - 1) {
-      ruleKey = self.matchAt(tokenIdx)
+      ruleKey = self.matchAt(tokenIdx, tokens)
       rule = self.rules[ruleKey]
       tokensMatched = rule.tokens
       self._ruleKeys.push(ruleKey)
@@ -294,10 +282,12 @@ module.exports = class GraphParser {
             if (self.matchTokens(
               tokenIdx - onmatch.prevClasses.length,
               onmatch.prevClasses,
+              tokens,
               true, false, true
             ) && self.matchTokens(
                 tokenIdx,
                 onmatch.nextClasses,
+                tokens,
                 false, true, true
               )) {
               output += onmatch.production